requirements.txt
flask
Flask-Cors==3.0.8
python-telegram-bot
gunicorn
firebase_admin
Pillow
google-cloud-vision